% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{abonnementRygg}
\alias{abonnementRygg}
\title{Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre}
\usage{
abonnementRygg(
  rnwFil,
  brukernavn = "tullebukk",
  reshID = 0,
  datoFra = Sys.Date() - 180,
  datoTil = Sys.Date()
)
}
\arguments{
\item{rnwFil}{Navn på fila som skal kjøres. Angis MED filending (\emph{dvs "filnavn.Rnw"})}

\item{reshID}{Aktuell reshid}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre
}
