% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{dataTilOffVisning}
\alias{dataTilOffVisning}
\title{Generere data til SKDEs interaktive nettsider
ODI er besvart ett år etter operasjon og resultatet vises for BESVARELSESÅR (Tenk over om skal filtrere på skjemadato eller ett år etter operasjonsdato.
Det siste er kanskje best basert på konsistens og mulighet for sammenligning med resultater som har utvalg på operasjonsdato.)}
\usage{
dataTilOffVisning(
  RegData = RegData,
  valgtVar,
  aar = 0,
  ktr = 0,
  indID = "indDummy",
  slaaSmToAar = 0,
  hovedkat = 99,
  hastegrad = 99,
  tidlOp = "",
  filUt = "dummy"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{\itemize{
\item beinsmLavPre, peropKompDura, sympVarighUtstr, p.t. 10 kvalitetsind.
}}

\item{aar}{\itemize{
\item Operasjonsår. Kan velge flere
}}

\item{indID}{indikator-id, eks. 'ind1', 'ind2', osv.}

\item{slaaSmToAar}{0:nei (standard), 1:ja. Slår sammen resultater for to og to år, glidende. Dvs. 2021 viser resultat fra alle operasjoner i
2020 og 2021, mens 2020 viser for 2020 og 2019. De fleste indikatorer vises for to år (slaaSmToAar=1). I praksis dupliseres data for hvert år.
Dette medfører at data må leveres aggregert ellers vil det bli tull med antall for hele landet.}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Andre inngrep',
\item 1:'Prolapskirurgi',
\item 2:'Midtlinjebevarende dekompr.',
\item 3:'Laminektomi',
\item 4:'Eksp. intersp implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Osteotomi, deformitet',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Skiveprotese',
\item 9:'Spinal stenose',
\item 10:'Degen. spondylolistese'
\item 99: Alle
}}

\item{hastegrad}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig}

\item{tidlOp}{tidligere operert? 1: samme nivå, 2: annet nivå, 3: annet og sm. nivå', 4: Primæroperasjon}

\item{filUt}{tilnavn for utdatatabell. Hvis ikke angitt, lastes ikke fil ned}
}
\value{
Datafil til Resultatportalen
}
\description{
Generere data til SKDEs interaktive nettsider
ODI er besvart ett år etter operasjon og resultatet vises for BESVARELSESÅR (Tenk over om skal filtrere på skjemadato eller ett år etter operasjonsdato.
Det siste er kanskje best basert på konsistens og mulighet for sammenligning med resultater som har utvalg på operasjonsdato.)
}
