% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{dataTilResPort}
\alias{dataTilResPort}
\title{Generere data til Resultatportalen}
\usage{
dataTilResPort(
  RegData = RegData,
  valgtVar,
  datoFra = "2011-01-01",
  aar = 0,
  hovedkat = 99,
  hastegrad = 99,
  tidlOp = "",
  filUt = "dummy"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{\itemize{
\item beinsmLavPre, peropKompDura, sympVarighUtstr
}}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{aar}{\itemize{
\item Operasjonsår. Kan velge flere
}}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Andre inngrep',
\item 1:'Prolapskirurgi',
\item 2:'Midtlinjebevarende dekompr.',
\item 3:'Laminektomi',
\item 4:'Eksp. intersp implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Osteotomi, deformitet',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Skiveprotese',
\item 9:'Spinal stenose',
\item 10:'Degen. spondylolistese'
\item 99: Alle
}}

\item{hastegrad}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig}

\item{tidlOp}{tidligere operert? 1: samme nivå, 2: annet nivå, 3: annet og sm. nivå', 4: Primæroperasjon}

\item{filUt}{tilnavn for utdatatabell (fjern?)}
}
\value{
Datafil til Resultatportalen
}
\description{
Generere data til Resultatportalen
}
