% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kategoriserInngrep.R
\name{defProSS}
\alias{defProSS}
\title{Definisjon av diagnosegrupper prolaps og spinal stenose, V3}
\usage{
defProSS(RegDataV3)
}
\arguments{
\item{RegDataV3}{}
}
\value{
Variablene LSSopr, ProlapsDekr, ProlapsFusjonert, ProlapsoprAlle
legges til opprinnelig dataramme
}
\description{
Definisjon av diagnosegrupper prolaps og spinal stenose, V3
}
