% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{delTekst}
\alias{delTekst}
\title{Automatisk linjebryting av lange tekstetiketter}
\usage{
delTekst(x, len)
}
\arguments{
\item{x}{En tekststreng eller vektor av tekststrenger}

\item{len}{Lengden strengen skal brytes ved}
}
\value{
automatisk linjebryting
}
\description{
Automatisk linjebryting av lange tekstetiketter
}
