% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{finnReoperasjoner}
\alias{finnReoperasjoner}
\title{Identifisere reoperasjoner
Legger på operasjonsnummer(OpNr), tid til neste operasjon (DagerNesteOp)
og identifiserer reoperasjoner innen 90 dager (Reop)}
\usage{
finnReoperasjoner(RegData)
}
\arguments{
\item{RegData}{}
}
\value{
Legger på operasjonsnummer, tid til neste og identifiserer reoperasjoner
}
\description{
Identifisere reoperasjoner
Legger på operasjonsnummer(OpNr), tid til neste operasjon (DagerNesteOp)
og identifiserer reoperasjoner innen 90 dager (Reop)
}
