% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{forsinketReg}
\alias{forsinketReg}
\title{Beregn andel registrereinger som er ferdigstilt viss lang tid etter operasjon.}
\usage{
forsinketReg(RegData, fraDato, tilDato, forsinkelse, reshID)
}
\arguments{
\item{RegData}{dataramme med reshID, FirstTimeClosed og operasjonsdato}

\item{fraDato}{startdato for perioden en ønsker å se på}

\item{tilDato}{sluttdato for perioden en ønsker å se på}

\item{forsinkelse}{minste antall dager fra operasjon til ferdigstillelse}

\item{reshID}{Avdelingas reshID. Benyttes til å filtrere.}
}
\value{

}
\description{
Forsinkelse er strengt tatt registrering og ferdigstillelse etter UtskrivelseDato,
men mht hva som er for sent for å sende ut oppfølgingsskjema, regnes forsinkelse
som tid etter operasjon
}
