% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHentData.R
\name{hentDataTabellV3}
\alias{hentDataTabellV3}
\title{Hent datatabell fra ngers database}
\usage{
hentDataTabellV3(
  tabellnavn = "surgeonform",
  qVar = "*",
  datoFra = "2019-01-01",
  datoTil = Sys.Date(),
  egneVarNavn = 1
)
}
\arguments{
\item{tabellnavn}{Navn på tabell som skal lastes inn.}

\item{egneVarNavn}{0 - Qreg-navn benyttes.
1 - selvvalgte navn fra Friendlyvar benyttes
Bare ferdigstilte (status=1) legeskjema og pasientskjema overføres}
}
\description{
Hent datatabell fra ngers database
}
