% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{henteSamlerapporter}
\alias{henteSamlerapporter}
\title{Funksjon som produserer rapporten som skal lastes ned av mottager.}
\usage{
henteSamlerapporter(
  filnavn,
  rnwFil,
  reshID = 0,
  datoFra = Sys.Date() - 180,
  datoTil = Sys.Date()
)
}
\arguments{
\item{filnavn}{brukes av downloadHandler}

\item{rnwFil}{Navn på fila som skal kjøres. Angis uten ending, dvs. (\emph{ uten ".Rnw"})}

\item{reshID}{Brukerens reshid}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}
}
\value{
Filsti til pdf-rapporten.
}
\description{
Funksjon som produserer rapporten som skal lastes ned av mottager.
}
