% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kategoriserInngrep.R
\name{kategoriserInngrep}
\alias{kategoriserInngrep}
\title{Kategorisere inngrep og hovedinngrep av operasjoner}
\usage{
kategoriserInngrep(RegData)
}
\arguments{
\item{RegData}{}
}
\value{
Variabler for inngrepskategorisering lagt til i RegData
}
\description{
Kategorisere inngrep og hovedinngrep av operasjoner
}
