% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{kjorRyggApp}
\alias{kjorRyggApp}
\title{Kjør Shiny Application}
\usage{
kjorRyggApp(browser = FALSE, logAsJson = FALSE)
}
\arguments{
\item{browser}{App åpner i browser}

\item{logAsJson}{Logg i json-format}
}
\value{
Et objekt som representerer Rygg-app'en
}
\description{
Kjør Shiny Application
}
