% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{lageTulleData}
\alias{lageTulleData}
\title{Lage tulledata (simulerte data)}
\usage{
lageTulleData(RegData, varBort = "", antSh = 26, antObs = 20000)
}
\arguments{
\item{RegData}{Ekte data som skal rotes til}

\item{varBort}{variable som finnes i begge filer}

\item{antSh}{antall sykehus}

\item{antObs}{antall observasjoner}
}
\description{
Lage tulledata (simulerte data)
}
