% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{lagTabavFig}
\alias{lagTabavFig}
\title{Vise figurdata som tabell}
\usage{
lagTabavFig(UtDataFraFig, figurtype = "andeler")
}
\arguments{
\item{UtDataFraFig}{beregnede og definerte verdier ut fra valgt variabel. Standard utdata fra figurberegningsfunksjonen}

\item{figurtype}{andeler, andelGrVar, andelTid, gjsnGrVar, gjsnTid}
}
\description{
Vise figurdata som tabell
}
