% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{lagTabavFigGjsnGrVar}
\alias{lagTabavFigGjsnGrVar}
\title{Vise figurdata som tabell, sentralmål per sykshus}
\usage{
lagTabavFigGjsnGrVar(UtDataFraFig)
}
\arguments{
\item{UtDataFraFig}{}
}
\description{
Vise figurdata som tabell, sentralmål per sykshus
}
