% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHentData.R
\name{mappingEgneNavn}
\alias{mappingEgneNavn}
\title{Endre variabelnavn/kolonnenavn til selvvalgte navn}
\usage{
mappingEgneNavn(tabell, tabType)
}
\arguments{
\item{tabell}{datatabellnavn i databasen}

\item{tabType}{REGISTRATION_TYPE}
}
\value{
tabell med selvvalgte variabelnavn spesifisert i friendlyvar. Intern funksjon
}
\description{
Endre variabelnavn/kolonnenavn til selvvalgte navn
}
