% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PverdiAndelsDiff.R
\name{PverdiAndelsDiff}
\alias{PverdiAndelsDiff}
\title{Beregn p-verdi, justert for multippeltesting, ved sammenligning av to og to andeler}
\usage{
PverdiAndelsDiff(n, N, konfnivaa = 0.95, justMetode = "fdr")
}
\arguments{
\item{n}{\itemize{
\item Antall tellende hendelser, skalar eller vektor
}}

\item{N}{\itemize{
\item Antall observasjoner, skalar eller vektor av samme lengde som n
}}

\item{konfnivaa}{\itemize{
\item Konfidensnivået til test (standard 0.95)
}}

\item{justMetode}{\itemize{
\item justeringsmetode, standard: 'fdr'. Mulige valg:
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr", "none"
}}
}
\value{
Justerte p-verdier for test av endring i andeler
}
\description{
Denne funksjonen benytter prop.test for å beregne p-verdi for at to andeler er like
og justerer p-verdiene ved p.adjust
}
