% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggV2V3_Torkil.R
\name{RyggDataV2V3}
\alias{RyggDataV2V3}
\title{Henter data registrert for Degenerativ Rygg}
\usage{
RyggDataV2V3()
}
\value{
RegData, dataramme med data f.o.m. 2007.
}
\description{
Henter data for Degenerativ Rygg og kobler samme versjon 2 og versjon 3.
Registeret ønsker også en versjon hvor variabler som bare er i versjon 2 er med i det
felles uttrekket.
}
