% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigAndeler.R
\name{RyggFigAndeler}
\alias{RyggFigAndeler}
\title{Søylediagram som viser andeler av ulike variabler}
\usage{
RyggFigAndeler(
  RegData,
  valgtVar,
  datoFra = "2007-01-01",
  datoTil = "2999-12-31",
  aar = 0,
  hentData = 0,
  preprosess = 1,
  minald = 0,
  maxald = 130,
  erMann = "",
  hovedkat = 99,
  hastegrad = 99,
  tidlOp = "",
  ktr = 0,
  tittelMed = 1,
  outfile = "",
  reshID = 0,
  enhetsUtvalg = 0,
  lagFig = 1
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{aar}{\itemize{
\item Operasjonsår. Kan velge flere
}}

\item{hentData}{Gjøre spørring mot database? 0: Nei, bruke RegData (Standard), 1: Ja}

\item{preprosess}{Preprosesser data 0: Nei, 1: Ja (Standard)}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Operasjonskategori: "ukjent"',    #hkat=0
\item 1:'Prolapskirurgi',
\item 2:'Foramenotomi',
\item 3:'Laminektomi',
\item 4:'Interspinøst implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Skiveprotese',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Spinal stenose',
\item 9:'Degen. spondylolistese'
\item 99: Alle
}}

\item{hastegrad}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig}

\item{tidlOp}{tidligere operert? 1: samme nivå, 2: annet nivå, 3: annet og sm. nivå', 4: Primæroperasjon}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{tittel}{Vise tittel i figuren eller ikke (0/1). standard:1
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Denne funksjonen lager et horisontalt eller vertikalt søylediagram som viser andeler (fordeling)
av valgt variabel filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
\itemize{
\item Alder: Aldersfordeling
\item AntDagerInnl: Liggetid
\item Antibiotika: Er det gitt antibiotikaprofylakse?
\item AntNivOpr: Antall nivå operert'
\item ArbstatusPre: Arbeidsstatus før operasjon'
\item -Arbstatus3mnd: Arbeidsstatus 3 mnd. etter operasjon
\item -Arbstatus12mnd: Arbeidsstatus 12 mnd. etter operasjon
\item ASA: ASA-grad
\item BMI: Pasientenes BMI (Body Mass Index)
\item EqangstPre: Helsetilstand: Angst
\item EqgangePre: Helsetilstand: Gange
\item ErstatningPre: Har pasienten søkt erstatning?
\item -Fornoyd3mnd: Fornøydhet 3 mnd etter operasjon
\item -Fornoyd12mnd: Fornøydhet 12 mnd etter operasjon
\item HovedInngrep: Hovedinngrep
\item Komorbiditet: Komorbiditet
\item KomplPer: Peroperative komplikasjoner
\item KomplPost: Pasientrapporterte komplikasjoner
\item Liggedogn: Liggetid ved operasjon
\item Morsmal: Morsmål
\item -Nytte3mnd: Hvilken nytte har du hatt av operasjonen? (svar 3 måneder etter)
\item -Nytte12mnd: Hvilken nytte har du hatt av operasjonen? (svar 12 måneder etter)
\item OpInd: Operasjonsindikasjon
\item OpIndPareseGrad: Operasjonsindikasjon, paresegrad
\item OpIndSmeType: Operasjonsindikasjon, smertetype
\item OpKat: Operasjonskategori
\item RadUnders: Radiologisk undersøkelse
\item Roker: Røyker du?
\item Saardren: Sårdren
\item SivilStatus: Sivilstatus
\item SmHyppPre: Hyppighet av smertestillende før operasjonen
\item SmStiPre: Bruk av smertestillende før operasjonen
\item SymptVarighRyggHof: Varighet av rygg-/hoftesmerter
\item SympVarighUtstr: Varighet av utstrålende smerter
\item TidlOpr: Tidligere ryggoperert?
\item TidlOprAntall: Antall tidligere operasjoner
\item UforetrygdPre: Har pasienten søkt uføretrygd?
\item Underkat: Fordeling av inngrepstyper. NB: hovedkategori MÅ velges
\item Utd: Høyeste fullførte utdanning
}
Argumentet \emph{hovedkat} har følgende valgmuligheter:
\itemize{
\item 0: Annet
\item 1: Prolaps
\item 2: Foramenotomi
\item 3: Laminektomi
\item 4: Eksp. intraspin. impl.
\item 5: Fusjon
\item 6: Skiveprotese
\item 7: Fjerning/revisjon
}
Velges ingen av disse, vil alle data vises.
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
\itemize{
\item 0: Hele landet
\item 1: Egen enhet mot resten av landet (Standard)
\item 2: Egen enhet
\item 3: Egen enhet mot egen sykehustype
\item 4: Egen sykehustype
\item 5: Egen sykehustype mot resten av landet
\item 6: Egen enhet mot egen region
\item 7: Egen region
\item 8: Egen region mot resten
}
}
