% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigAndelStabelTid.R
\name{RyggFigAndelStabelTid}
\alias{RyggFigAndelStabelTid}
\title{Søylediagram med en søyle for hvert år hvor variabelens kategorier er stablet oppå hverandre.}
\usage{
RyggFigAndelStabelTid(RegData, outfile, valgtVar, hovedkat = 99,
  preprosess = 0, hentData = 0, minald = 0, maxald = 130,
  erMann = 99, ktr = 1, tidlOp = 0, aar = 0, enhetsUtvalg = 0,
  tittel = 1, reshID = 0)
}
\arguments{
\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Operasjonskategori: "ukjent"',    #hkat=0
\item 1:'Prolapskirurgi',
\item 2:'Foramenotomi',
\item 3:'Laminektomi',
\item 4:'Interspinøst implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Skiveprotese',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Spinal stenose',
\item 9:'Degen. spondylolistese'
\item 99: Alle
}}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{ktr}{Oppfølging nr. 1 - 3mnd kontroll, 2 - 12 mnd kontroll}

\item{tidlOp}{tidligere operert? 1: samme nivå, 2: annet nivå, 3: annet og sm. nivå', 4: Primæroperasjon}

\item{aar}{\itemize{
\item Operasjonsår
}}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
Videreutvikling: Ikke vis søyler (år) som har < 10 registreringer.
}
\description{
Denne funksjonen lager et vertikalt søylediagram hvor søylene er delt inn i variabelens kategorier. Søylene
har følgelig alle høyde 1.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
\itemize{
\item Fornoyd: Hvilken nytte mener du at du har hatt av operasjonen?
\item Nytte: Hvor stor nytte pasienten har hatt av operasjonen
\item TidlOp: Tidligere operasjon
}
Argumentet \emph{hovedkat} har følgende valgmuligheter:
\itemize{
\item 0: Annet
\item 1: Prolaps
\item 2: Foramenotomi
\item 3: Laminektomi
\item 4: Eksp. intraspin. impl.
\item 5: Fusjon
\item 6: Skiveprotese
\item 7: Fjerning/revisjon
}
Velges ingen av disse, vil alle data vises.
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
\itemize{
\item 0: Hele landet
\item 1: Egen enhet mot resten av landet (Standard)
\item 2: Egen enhet
\item 3: Egen enhet mot egen sykehustype
\item 4: Egen sykehustype
\item 5: Egen sykehustype mot resten av landet
\item 6: Egen enhet mot egen region
\item 7: Egen region
\item 8: Egen region mot resten
}
}
