% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggFigGrupperLinjeTid.R
\name{RyggFigGrupperLinjelTid}
\alias{RyggFigGrupperLinjelTid}
\title{Tidstrend av andel opphold}
\usage{
RyggFigGrupperLinjelTid(
  RegData,
  valgtVar = "gruppeAndel",
  gr1 = "",
  gr2 = "",
  gr3 = "",
  datoFra = "2011-01-01",
  datoTil = Sys.Date(),
  aar = 0,
  tidsenhet = "Aar",
  hovedkat = 99,
  ktr = 0,
  tidlOp = 99,
  tittel = 1,
  minald = 0,
  maxald = 130,
  erMann = 99,
  reshID = 0,
  outfile = "",
  hastegrad = 99,
  enhetsUtvalg = 0,
  preprosess = 1,
  hentData = 0,
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{aar}{\itemize{
\item Operasjonsår. Kan velge flere
}}

\item{tidsenhet}{Oppløsning på tidsaksen. Verdier: 'Aar' (standard), 'Halvaar', 'Kvartal','Mnd'}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Andre inngrep',
\item 1:'Prolapskirurgi',
\item 2:'Midtlinjebevarende dekompr.',
\item 3:'Laminektomi',
\item 4:'Eksp. intersp implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Osteotomi, deformitet',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Skiveprotese',
\item 9:'Spinal stenose',
\item 10:'Degen. spondylolistese'
\item 99: Alle
}}

\item{tidlOp}{tidligere operert? 1: samme nivå, 2: annet nivå, 3: annet og sm. nivå', 4: Primæroperasjon}

\item{tittel}{Vise tittel i figuren eller ikke (0/1). standard:1
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 110)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives}

\item{hastegrad}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig}

\item{enhetsUtvalg}{0: hele landet, 2: eget}

\item{preprosess}{Preprosesser data 0: Nei, 1: Ja (Standard)}

\item{hentData}{Gjøre spørring mot database? 0: Nei, bruke RegData (Standard), 1: Ja}
}
\value{
Figur som viser tidstrend, enten andel for hver av valgte grupper eller for en valgt responsvariabel
}
\description{
Funksjonen lager linjediagram som viser utvikling over tid av valgt variabel
for valgte grupper, filtrert på de utvalg som er gjort.
Per nå (23.mai 2024) må funksjonen kjøres manuelt. Ønsker å gjøre den mer generisk. Avventer registerets behov.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
\itemize{
\item xx
\item OswEndr30pst12mnd: Forbedring av Oswestry-skår > 30 prosent
}
}
