% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHentData.R
\name{RyggHentRegData}
\alias{RyggHentRegData}
\title{Henter Rygg-tabeller og kobler sammen}
\usage{
RyggHentRegData(datoFra = "2013-01-01", datoTil = Sys.Date(), medOppf = 1, ...)
}
\arguments{
\item{medPROM:}{koble på RAND og TSS2-variabler}

\item{alleData}{1- alle variabler med, 0 - utvalgte variabler med}
}
\value{
RegData data frame
}
\description{
Henter Rygg-tabeller og kobler sammen
}
