% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggPreprosesser.R
\name{RyggPreprosess2}
\alias{RyggPreprosess2}
\title{Preprosesser data fra Degenerativ Rygg}
\usage{
RyggPreprosess2(RegData = RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\description{
Denne funksjonen definerer og formaterer variabler
}
