% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggRegDataOppdateres.R
\name{RyggRegDataSQL}
\alias{RyggRegDataSQL}
\title{Henter data registrert for Degenerativ Rygg}
\usage{
RyggRegDataSQL()
}
\value{
RegData data frame
}
\description{
Henter data for Degenerativ Rygg fra "staging" (?)
}
