% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggRegDataSQL.R
\name{RyggRegDataSQLgml}
\alias{RyggRegDataSQLgml}
\title{Henter data registrert for Degenerativ Rygg}
\usage{
RyggRegDataSQLgml(datoFra = "2007-01-01", datoTil = "2099-01-01", alle = 1)
}
\value{
RegData data frame
}
\description{
Henter data for Degenerativ Rygg V3, dvs. fra ca 1.jan. 2019
}
