% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggUtvalgEnh.R
\name{RyggUtvalgEnh}
\alias{RyggUtvalgEnh}
\title{Title}
\usage{
RyggUtvalgEnh(
  RegData,
  datoFra = "2009-01-01",
  datoTil = "3000-01-01",
  minald = 0,
  maxald = 130,
  erMann = "",
  hovedkat = 99,
  aar = 0,
  tidlOp = 99,
  hastegrad = 99,
  enhetsUtvalg = 0,
  reshID = 0,
  fargepalett = "BlaaOff"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Operasjonskategori: "ukjent"',    #hkat=0
\item 1:'Prolapskirurgi',
\item 2:'Foramenotomi',
\item 3:'Laminektomi',
\item 4:'Interspinøst implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Skiveprotese',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Spinal stenose',
\item 9:'Degen. spondylolistese'
\item 99: Alle
}}

\item{aar}{\itemize{
\item Operasjonsår. Kan velge flere
}}

\item{tidlOp}{tidligere operert? 1: samme nivå, 2: annet nivå, 3: annet og sm. nivå', 4: Primæroperasjon}

\item{hastegrad}{Hastegrad av operasjon 1: Elektivt, 2: Akutt, 3: Halvøyeblikkelig}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir}

\item{fargepalett}{Velge fargepalett, standard:BlaaOff ("offentliggjøringsfargene")}
}
\value{
Returnerer filtrert versjon av RegData
}
\description{
Funksjon som gjør utvalg i datagrunnlaget til registreringene for Nakke
}
\details{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og utvalgsteksten.
Gjøre ingen utvalg, vil alle data vises.

Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
\itemize{
\item 0: Hele landet (Standard)
\item 1: Egen enhet mot resten av landet
\item 2: Egen enhet
\item 3: Egen enhet mot egen sykehustype
\item 4: Egen sykehustype
\item 5: Egen sykehustype mot resten av landet
\item 6: Egen enhet mot egen region
\item 7: Egen region
\item 8: Egen region mot resten
}
}
