% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggVarTilrettelegg.R
\name{RyggVarTilrettelegg}
\alias{RyggVarTilrettelegg}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
RyggVarTilrettelegg(
  RegData = NULL,
  valgtVar,
  ktr = 0,
  datoFra = "2007-01-01",
  datoTil = Sys.Date(),
  hovedkat = 99,
  figurtype = "andeler"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste operasjonsdato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Andre inngrep',
\item 1:'Prolapskirurgi',
\item 2:'Midtlinjebevarende dekompr.',
\item 3:'Laminektomi',
\item 4:'Eksp. intersp implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Osteotomi, deformitet',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Skiveprotese',
\item 9:'Spinal stenose',
\item 10:'Degen. spondylolistese'
\item 99: Alle
}}

\item{figurtype}{Hvilken figurtype det skal tilrettelegges variable for:
'andeler', 'andelGrVar', 'andelTid', 'gjsnGrVar', 'gjsnTid'}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk.
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt.
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen.
Her kan mye hentes til analysebok
Når ei fordeling består av flere variable, settes flerevar=1. Vi omdefinerer variablene slik at
alle gyldige registreringer (dvs. alle registreringer som skal telles med) er 0 eller 1.
De som har oppfylt spørsmålet
er 1, mens ugyldige registreringer er NA. Det betyr at hvis vi skal ta bort registreringer
som i kategorier av typen "Ukjent" kodes disse som NA, mens hvis de skal være med kodes de som 0.
Vi sender tilbake alle variable som indikatorvariable, dvs. med 0,1,NA
}
