% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appRygg.R
\name{server_rygg}
\alias{server_rygg}
\title{Server-del til appen}
\usage{
server_rygg(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
Server-delen til Rygg-appen
}
\description{
Server-del til appen
}
