% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{SorterOgNavngiTidsEnhet}
\alias{SorterOgNavngiTidsEnhet}
\title{Hjelpefunksjoner. Group of functions page title}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)
}
\arguments{
\item{RegData}{registerdata}

\item{tidsenhet}{'AAr', 'Halvaar', 'Kvartal' eller 'Mnd'}

\item{tab}{? hmm...}
}
\description{
Fil med div hjelpefunksjoner.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.

Tilrettelegge tidsenhetvariabel:
}
