% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKjustering.R
\name{StandAlderKjonn}
\alias{StandAlderKjonn}
\title{Juster datamaterialet for alder og kjønn}
\usage{
StandAlderKjonn(RegData, stdPop, antAldgr, katVariable)
}
\arguments{
\item{RegData}{Ei dataramme som minimum må inneholde variabelen det skal beregnes alders- og
kjønnsstandardiserte rater for, samt alder, kjønn og bo-områder.}

\item{stdPop}{standardpopulasjon: 'Norge'-juster mot befolkningfil fra SSB.
'Register'-juster mot alderssammensetning i registeret}

\item{antAldgr}{antall aldersgrupper man ønsker å benytte for justering mot alderssammensetning i befolkninga}

\item{katVariable}{grupperingsvariable, eks. 'Aar' eller 'ShNavn'}
}
\value{
Funksjon som alders- og kjønnsstandardiserer data basert på en valgt standardpopulasjon.
}
\description{
Standardiserer beregning av en "andels"variabel mot variablene  Alder (år) og ErMann (0/1) fra valgt standardpopulasjon
}
