% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{tabAntOpphSh5Aar}
\alias{tabAntOpphSh5Aar}
\title{Tabell som viser antall opphold per sykehus og år, siste 5 år.}
\usage{
tabAntOpphSh5Aar(RegData, datoTil = Sys.Date())
}
\arguments{
\item{RegData}{Registerdata}

\item{datoTil}{sluttdato for visningsperiode}
}
\value{
Antall opphold per sykehus og år, siste 5 år
}
\description{
Tabell som viser antall opphold per sykehus og år, siste 5 år.
}
