% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{tabAntOpphShMnd}
\alias{tabAntOpphShMnd}
\title{Tabell som viser antall operasjoner per måned og sykehus siste x måneder}
\usage{
tabAntOpphShMnd(RegData, datoTil = Sys.Date(), antMnd = 6, reshID = 0)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}
}
\value{
Antall operasjoner per måned og sykehus.
}
\description{
Hvis reshID oppgis kommer månedlig oversikt kun for det aktuelle sykehuset. RegData må inneholde InnDato og Aar.
Tabellen returneres som en xtable
}
