% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{tabAntSkjema}
\alias{tabAntSkjema}
\title{Hvor mange skjema av hver type (lege, pasient, oppfølging)}
\usage{
tabAntSkjema(RegData, datoFra = "2019-01-01", datoTil = Sys.Date())
}
\arguments{
\item{RegData}{Tabellen skjemaoversikt}

\item{datoFra}{angi start for tidsperioden}

\item{datoTil}{angi slutt for tidsperioden}
}
\description{
Tabell som viser oversikt over antall skjema av hver type. Kan velge kladd/ferdigstilt
}
