% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{tabNokkeltall}
\alias{tabNokkeltall}
\title{Nøkkeltall ( div oversiktstall)}
\usage{
tabNokkeltall(
  RegData,
  tidsenhet = "Mnd",
  datoTil = Sys.Date(),
  enhetsUtvalg = 2,
  reshID = 0
)
}
\arguments{
\item{RegData}{}

\item{reshID}{}
}
\description{
Nøkkeltall ( div oversiktstall)
}
