% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggPreprosesser.R
\name{tilpassV2data}
\alias{tilpassV2data}
\title{Justere V2-data}
\usage{
tilpassV2data(RegDataV2)
}
\arguments{
\item{RegDataV2}{}
}
\value{
dataramme med tilpassede V2-data
}
\description{
Justere V2-data
}
