% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggPreprosesser.R
\name{tilpassV3data}
\alias{tilpassV3data}
\title{Justere V3-data}
\usage{
tilpassV3data(RegDataV3)
}
\arguments{
\item{RegDataV3}{}
}
\value{
dataramme med tilpassede V3-data
}
\description{
Justere V3-data
}
