% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{tilretteleggDataDumper}
\alias{tilretteleggDataDumper}
\title{Lage entydig PID
Lager entydig PID basert på personnummer. Alle PID i V2 har ending suffiks V2. Pasienter i V3 får PID fra V2 hvis personnummeret
deres finnes i begge versjoner. NB: For at dette skal bli riktig er vi avhengig av kilde med oppdatert koblingsfil (PID-personnummer)}
\usage{
tilretteleggDataDumper(
  RegData,
  datoFra = "2000-01-01",
  datoTil = Sys.Date(),
  reshID = 0
)
}
\arguments{
\item{datoFra}{startdato for filtrering}

\item{datoTil}{sluttdato for filtrering}

\item{reshID}{hvilken resh det evt. skal filtreres på}

\item{data}{tilnavn for utdatatabell. Hvis ikke angitt, lastes ikke fil ned}
}
\value{
Datafil med entydige PID
}
\description{
Lage entydig PID
Lager entydig PID basert på personnummer. Alle PID i V2 har ending suffiks V2. Pasienter i V3 får PID fra V2 hvis personnummeret
deres finnes i begge versjoner. NB: For at dette skal bli riktig er vi avhengig av kilde med oppdatert koblingsfil (PID-personnummer)
}
