% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{kjorRyggApp}
\alias{kjorRyggApp}
\title{Hjelpefunksjoner. Group of functions page title}
\usage{
kjorRyggApp()
}
\value{
Et objekt som representerer den Rygg-app'en
}
\description{
Fil med div hjelpefunksjoner.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.

Kjør Shiny Application
}
