% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggPreprosesser.R
\name{RyggPreprosess}
\alias{RyggPreprosess}
\title{Preprosesser data fra Degenerativ Rygg}
\usage{
RyggPreprosess(RegData = RegData)
}
\description{
Denne funksjonen definerer og formaterer variabler
}
