% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggVarTilrettelegg.R
\name{RyggVarTilrettelegg}
\alias{RyggVarTilrettelegg}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
RyggVarTilrettelegg(RegData = NULL, valgtVar, ktr = 0, hovedkat = 99,
  figurtype = "andeler")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{hovedkat}{Hvilken type hovedinngrep, numerisk 0-9, standard: 99, dvs. alle. Valgmuligheter:
\itemize{
\item 0:'Operasjonskategori: "ukjent"',    #hkat=0
\item 1:'Prolapskirurgi',
\item 2:'Foramenotomi',
\item 3:'Laminektomi',
\item 4:'Interspinøst implantat',
\item 5:'Fusjonskirurgi',
\item 6:'Skiveprotese',
\item 7:'Fjerning/rev. av implantat',
\item 8:'Spinal stenose',
\item 9:'Degen. spondylolistese'
\item 99: Alle
}}

\item{figurtype}{Hvilken figurtype det skal tilrettelegges variable for:
'andeler', 'andelGrVar', 'andelTid', 'gjsnGrVar', 'gjsnTid'}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk.
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt.
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen.
Her kan mye hentes til analysebok
}
