% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggHjelpefunksjoner.R
\name{SorterOgNavngiTidsEnhet}
\alias{SorterOgNavngiTidsEnhet}
\title{Tilrettelegge tidsenhetvariabel:}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)
}
\arguments{
\item{RegData}{registerdata}

\item{tidsenhet}{'AAr', 'Halvaar', 'Kvartal' eller 'Mnd'}

\item{tab}{? hmm...}
}
\description{
Tilrettelegge tidsenhetvariabel:
}
