% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{tabAntSkjema}
\alias{tabAntSkjema}
\title{Hvor mange skjema av hver type (lege, pasient, oppfølging)}
\usage{
tabAntSkjema(
  SkjemaOversikt,
  datoFra = "2019-01-01",
  datoTil = Sys.Date(),
  skjemastatus = 1
)
}
\arguments{
\item{SkjemaOversikt}{Tabellen skjemaoversikt}

\item{datoFra}{angi start for tidsperioden}

\item{datoTil}{angi slutt for tidsperioden}

\item{skjemastatus}{0: Kladd, 1:ferdigstilt}
}
\description{
Tabell som viser oversikt over antall skjema av hver type. Kan velge kladd/ferdigstilt
}
