% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RyggTabeller.R
\name{tabNokkeltall}
\alias{tabNokkeltall}
\title{Nøkkeltall ( div oversiktstall)}
\usage{
tabNokkeltall(
  RegData,
  tidsenhet = "Mnd",
  datoTil = Sys.Date(),
  enhetsUtvalg = 2,
  reshID = 0
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoTil}{Seneste operasjonsdato i utvalget (vises alltid i figuren).}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir}
}
\description{
Nøkkeltall ( div oversiktstall)
}
